ai_research_analyst/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
pydantic>=2.0.0
